fix import for data collector and add birthtime feature
