+++ data_collector/.gitattributes
