+++ data_collector/analysisScripts/analysisJH.R
