+++ data_collector/analysisScripts/analysisJH.bat
