+++ data_collector/analysisScripts/statAnalysis.R
