+++ data_collector/featureExtractors/Arc.py
@@ -2,8 +2,8 @@ import math
 from scipy.optimize import fsolve
 import os
 from FeatureExtractorAbstract import FeatureExtractorAbstract
-from helpers.config import PathConfig
-from helpers.distanceCalc import DistanceCalc
+from ..helpers.config import PathConfig
+from ..helpers.distanceCalc import DistanceCalc
 
 
 class Arc(FeatureExtractorAbstract):
