+++ data_collector/featureExtractors/FeatureExtractorAbstract.py
