+++ data_collector/featureExtractors/Gait.py
@@ -1,8 +1,8 @@
 import numpy as np
 import os
 from FeatureExtractorAbstract import FeatureExtractorAbstract
-from helpers.config import PathConfig
-from helpers.distanceCalc import DistanceCalc
+from ..helpers.config import PathConfig
+from ..helpers.distanceCalc import DistanceCalc
 
 
 class Gait(FeatureExtractorAbstract):
