+++ data_collector/featureExtractors/MutProbability.py
@@ -1,7 +1,7 @@
 import os
 from FeatureExtractorAbstract import FeatureExtractorAbstract
-from helpers.config import PathConfig
-from helpers.getVoxelData import VoxelData
+from ..helpers.config import PathConfig
+from ..helpers.getVoxelData import VoxelData
 
 
 class MutProbability(FeatureExtractorAbstract):
