+++ data_collector/featureExtractors/RelativeCellCountAlt.py
@@ -1,8 +1,8 @@
 import os
 from FeatureExtractorAbstract import FeatureExtractorAbstract
-from helpers.config import PathConfig
-from helpers.getVoxelData import VoxelData
-from helpers.getAltFile import GetAltFile
+from ..helpers.config import PathConfig
+from ..helpers.getVoxelData import VoxelData
+from ..helpers.getAltFile import GetAltFile
 
 
 class RelativeCellCountAlt(FeatureExtractorAbstract):
