+++ data_collector/featureExtractors/SizeOnAxis.py
@@ -1,8 +1,8 @@
 import os
 import types
 from FeatureExtractorAbstract import FeatureExtractorAbstract
-from helpers.config import PathConfig
-from helpers.getVoxelData import VoxelData
+from ..helpers.config import PathConfig
+from ..helpers.getVoxelData import VoxelData
 
 
 class SizeOnAxis(FeatureExtractorAbstract):
