+++ data_collector/getDistanceForTraceFile.py
