+++ data_collector/helpers/getDistanceForTraceFile.py
